+++ front-end/src/components/dashboard/dashboard.scss
@@ -1,5 +1,5 @@
 @import '../../../style/variables.scss';
 
 .dashboard {
-  padding-top: 100px;
+  padding-top: 75px;
 }
